Rework value assignment so it uses small builders
