+++ packages/auto_mappr/lib/src/builder/assignments/assignments.dart
@@ -1,0 +1,4 @@
+export 'assignment_builder_base.dart';
+export 'iterable_assignment_builder.dart';
+export 'map_assignment_builder.dart';
+export 'nested_object_assignment_builder.dart';
