+++ packages/auto_mappr/lib/src/builder/assignments/nested_object_mixin.dart
@@ -1,0 +1,135 @@
+import 'package:analyzer/dart/element/nullability_suffix.dart';
+import 'package:analyzer/dart/element/type.dart';
+import 'package:auto_mappr/src/builder/assignments/assignment_builder_base.dart';
+import 'package:auto_mappr/src/builder/methods/method_builder_base.dart';
+import 'package:auto_mappr/src/extensions/dart_type_extension.dart';
+import 'package:auto_mappr/src/models/models.dart';
+import 'package:build/build.dart';
+import 'package:code_builder/code_builder.dart';
+import 'package:source_gen/source_gen.dart';
+
+/// Separate logic as a mixin so we do not soil [AssignmentBuilderBase].
+mixin NestedObjectMixin on AssignmentBuilderBase {
+  /// Assigns nested object as either:
+  /// - default value
+  /// - call to already generated mapping between two types
+  ///
+  /// If [convertMethodArgument] is null, uses a tear off call instead.
+  Expression assignNestedObject({
+    required DartType source,
+    required DartType target,
+    required SourceAssignment assignment,
+    Expression? convertMethodArgument,
+    bool includeGenericTypes = false,
+  }) {
+    if (source.isSame(target)) {
+      return refer('model').property(assignment.sourceField!.displayName);
+    }
+
+    final nestedMapping = mapperConfig.findMapping(
+      source: source,
+      target: target,
+    );
+
+    if (nestedMapping == null) {
+      final targetTypeName = target.getDisplayStringWithLibraryAlias(
+        withNullability: true,
+        config: mapperConfig,
+      );
+      final sourceName = assignment.sourceField?.getDisplayString(withNullability: true);
+
+      if (target.nullabilitySuffix == NullabilitySuffix.question) {
+        log.warning("Can't find nested mapping '$assignment' but target is nullable. Setting null");
+
+        return literalNull;
+      }
+
+      throw InvalidGenerationSourceError(
+        'Trying to map nested object from "$assignment" but no mapping is configured.',
+        todo: 'Configure mapping from $sourceName to $targetTypeName',
+      );
+    }
+
+    final convertCallExpression = mappingCall(
+      nestedMapping: nestedMapping,
+      source: source,
+      target: target,
+      convertMethodArgument: convertMethodArgument,
+      includeGenericTypes: includeGenericTypes,
+    );
+
+    // If source == null and target not nullable -> use whenNullDefault if possible
+    final fieldMapping = mapping.tryGetFieldMapping(assignment.targetName);
+    if (source.nullabilitySuffix == NullabilitySuffix.question && (fieldMapping?.whenNullExpression != null)) {
+      // Generates code like:
+      //
+      // model.name == null
+      //     ? const Nested(
+      //         id: 123,
+      //         name: 'test',
+      //       )
+      //     : _map_NestedDto_To_Nested(model.name),
+      return refer('model').property(assignment.sourceField!.displayName).equalTo(literalNull).conditional(
+            fieldMapping!.whenNullExpression!,
+            convertCallExpression,
+          );
+    }
+
+    // Generates code like:
+    //
+    // `_map_NestedDto_To_Nested(model.name)`
+    return convertCallExpression;
+  }
+
+  /// Generates a mapping call `_mapAlphaDto_to_Alpha(convertMethodArgument)`.
+  /// When [convertMethodArgument] is null, then a tear off `_mapAlphaDto_to_Alpha` is generated.
+  ///
+  /// This function also marks nullable mapping to be generated
+  /// using the [usedNullableMethodCallback] callback.
+  Expression mappingCall({
+    required DartType source,
+    required DartType target,
+    required TypeMapping nestedMapping,
+    Expression? convertMethodArgument,
+    bool includeGenericTypes = false,
+  }) {
+    final targetNullable = target.nullabilitySuffix == NullabilitySuffix.question;
+
+    final useNullableMethod = targetNullable && !mapping.hasWhenNullDefault();
+
+    // When target is nullable, use nullable convert method.
+    // But use non-nullable when the mapping has default value.
+    //
+    // Otherwise use non-nullable.
+    final convertMethod = refer(
+      useNullableMethod
+          ? MethodBuilderBase.constructNullableConvertMethodName(
+              source: source,
+              target: target,
+              config: mapperConfig,
+            )
+          : MethodBuilderBase.constructConvertMethodName(
+              source: source,
+              target: target,
+              config: mapperConfig,
+            ),
+    );
+
+    if (useNullableMethod) {
+      usedNullableMethodCallback?.call(nestedMapping);
+    }
+
+    return convertMethodArgument == null
+        ? convertMethod
+        : convertMethod.call(
+            [convertMethodArgument],
+            {},
+            includeGenericTypes
+                ? [
+                    refer(source.getDisplayStringWithLibraryAlias(withNullability: true, config: mapperConfig)),
+                    refer(target.getDisplayStringWithLibraryAlias(withNullability: true, config: mapperConfig)),
+                  ]
+                : [],
+          );
+  }
+}
