+++ packages/auto_mappr/lib/src/builder/methods/callable_method.dart
@@ -5,6 +5,4 @@ abstract class CallableMethod {
   const CallableMethod();
 
   Expression methodCall({Map<String, Expression> namedArguments = const {}});
-
-  
 }
