+++ pubspec.yaml
@@ -2,7 +2,7 @@ name: auto_mappr_workspace
 publish_to: "none"
 
 environment:
-  sdk: ">=2.18.0 <4.0.0"
+  sdk: ^3.0.0
 
 dev_dependencies:
   melos: ^3.0.0
